Adding Messiaen's 7 Modes Of Limited transposition
I would like to suggest/discuss the adding of Messiaen's 7 "Modes Of Limited Transposition". The addition consists partly of adding a new name to an existing scale, but also adding new scales.

As for the naming I am unsure what would be the best. Currently "messiaen's mode of limited transposition #x" is used, which is exact, but also very long and uses an apostrophe, which might cause troubles. Personally, I would be fine with something like "messiaen #x" or "mode of limited transposition #x" (which would at least get rid of the potential apostrophe problem).

Source:
https://en.wikipedia.org/wiki/Mode_of_limited_transposition
https://es.wikipedia.org/wiki/Modos_de_transposici%C3%B3n_limitada
https://de.wikipedia.org/wiki/Modi_mit_begrenzten_Transpositionsm%C3%B6glichkeiten

Muchas gracias, Daniel!